Refactoring the weather enums into abstraction model class
